Move stream_anayzer to interpreter
